requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyYAML==3.12
redis==2.10.5
requests==2.18.2
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2
